Apply comments in PR
